feat(ts-sdk): Support web-content as a Manifestation type (#119)
EPMC has docmaps that use web-content as the type of a Manifestation. However, the sdk up until this point has only supported Manifestations with a type of web-page.

Note: web-content in the fabioverse is technically an Expression and not a Manifestation, so semantically, this is incorrect.

Therefore, we should attempt to deprecate the use of web-content in manifestations, but for the time being parse docmaps which contain it.