Fix Docmap Explorer in Firefox (#218)
The Docmap Explorer wasn't working in Firefox and Safari. This is because a CORS error that only happens in those browsers was causing the tabs containing the widget and the crossref demo not to be displayed. The CORS error happened when fetching data from crossref.

This PR comments out the request to crossref and the crossref tab in the explorer UI. We can bring those back once the CORS issue is fixed.

Other changes in this commit:

* The explorer's page title is now "Docmap Explorer" instead of "Docmap Demonstration"
* Adjust font styling for tooltips in the widget